requirements.txt
pytest==8.0.0
selenium==4.17.2
webdriver-manager==4.0.1
allure-pytest==2.13.2